+++ packages/framework/tests/Feature/ConfigurableFeaturesTest.php
@@ -64,6 +64,36 @@ class ConfigurableFeaturesTest extends TestCase
         $this->assertFalse(Features::hasSitemap());
     }
 
+    public function testHasRobotsTxtReturnsTrueByDefault()
+    {
+        $this->assertTrue(Features::hasRobotsTxt());
+    }
+
+    public function testHasRobotsTxtReturnsFalseWhenDisabledInConfig()
+    {
+        config(['hyde.robots.enabled' => false]);
+        $this->assertFalse(Features::hasRobotsTxt());
+    }
+
+    public function testHasLlmsTxtReturnsTrueByDefaultWhenHydeHasBaseUrl()
+    {
+        $this->withSiteUrl();
+        $this->assertTrue(Features::hasLlmsTxt());
+    }
+
+    public function testHasLlmsTxtReturnsFalseIfHydeDoesNotHaveBaseUrl()
+    {
+        config(['hyde.url' => '']);
+        $this->assertFalse(Features::hasLlmsTxt());
+    }
+
+    public function testHasLlmsTxtReturnsFalseWhenDisabledInConfig()
+    {
+        $this->withSiteUrl();
+        config(['hyde.llms.enabled' => false]);
+        $this->assertFalse(Features::hasLlmsTxt());
+    }
+
     public function testHasThemeToggleButtonsReturnsTrueWhenDarkmodeEnabledAndConfigTrue()
     {
         // Enable dark mode and set hyde.theme_toggle_buttons config option to true