+++ packages/framework/tests/Feature/RedirectTest.php
@@ -64,6 +64,14 @@ class RedirectTest extends TestCase
         Filesystem::unlink('_site/foo.html');
     }
 
+    public function testDottedRedirectPathKeepsItsExtension()
+    {
+        $redirect = new Redirect('legacy.json', 'new-location');
+
+        $this->assertSame('legacy.json', $redirect->getRouteKey());
+        $this->assertSame('legacy.json', $redirect->getOutputPath());
+    }
+
     public function testRedirectsCannotWriteOutsideTheBuildPipeline()
     {
         $this->assertFalse(method_exists(Redirect::class, 'create'));
@@ -73,5 +81,10 @@ class RedirectTest extends TestCase
     public function testRedirectsAreHiddenFromNavigation()
     {
         $this->assertFalse((new Redirect('foo', 'bar'))->showInNavigation());
+    }
+
+    public function testRedirectsAreHiddenFromSitemaps()
+    {
+        $this->assertFalse((new Redirect('foo', 'bar'))->showInSitemap());
     }
 }