+++ packages/framework/tests/Unit/GenerateBuildManifestTest.php
@@ -20,6 +20,8 @@ class GenerateBuildManifestTest extends UnitTestCase
 
     public function testActionGeneratesBuildManifest()
     {
+        self::mockConfig(['hyde.robots.enabled' => false, 'hyde.llms.enabled' => false]);
+
         Hyde::pages()->addPage(new DocumentationSearchIndex());
 
         (new GenerateBuildManifest())->handle();